pyproject.toml
[tool.poetry]
name = "cool"
version = "0.1.0"
description = ""
authors = ["d2verb"]

[tool.poetry.dependencies]
python = "^3.8"
pycryptodome = "^3.9.8"
gmpy2 = "^2.0.8"
requests = "^2.24.0"
fire = "^0.3.1"
termcolor = "^1.1.0"
pyelftools = "^0.26"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.8.3"
mypy = "^0.782"
isort = "^4.3.21"
bpython = "^0.19"

[tool.poetry.scripts]
ctf = "cool.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
